add explicit downcast

+++ tools/flang2/flang2exe/iliutil.cpp
@@ -2093,7 +2093,7 @@ reciprocal_division(int n, INT divisor, int sgnd)
   } else {
     INT64 twol;
     int shpre = 0;
-    INT64 one_64 = {0x0, 0x1}, divisor64 = {0x0, udiv};
+    INT64 one_64 = {0x0, 0x1}, divisor64 = {0x0, (int) udiv};
     const int l = lg(udiv);
 
     shf64(one_64, l, twol);
